+++ ostd/src/mm/mod.rs
@@ -26,7 +26,7 @@ use spin::Once;
 
 pub use self::{
     dma::{Daddr, DmaCoherent, DmaDirection, DmaStream, DmaStreamSlice, HasDaddr},
-    frame::{options::FrameAllocOptions, Frame, FrameVec, FrameVecIter, Segment},
+    frame::{options::FrameAllocOptions, Frame, Segment},
     io::{KernelSpace, UserSpace, VmIo, VmReader, VmWriter},
     page_prop::{CachePolicy, PageFlags, PageProperty},
     vm_space::VmSpace,
